pyproject.toml
[project]
name = "transcriber"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.32.3",
    "flask",
    "openai",
    "python-dotenv"
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"
